+++ packages/eslint-config-sslcom/base.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['eslint-config-sslcom-base'].map(require.resolve),
+  rules: {},
+};