+++ packages/eslint-config-airbnb-sslcom/CHANGELOG.md
@@ -66,4 +66,4 @@
 
 1.0.0 / 2016-04-19
 ==================
- - Initial commmit; moved content over from `eslint-config-airbnb` package.
+ - Initial commmit; moved content over from `eslint-config-sslcom` package.